COMMENT: sevice & controller 추가할 기능 회의

+++ trip-one-zido-api/src/main/java/com/example/triponezidoapi/controller/CommunityApiController.java
@@ -79,7 +79,7 @@ public class CommunityApiController {
 
             @PathVariable
             @Parameter(description = "페이지 번호")
-            int page
+            long page
     ){
         return communityService.getCommunityListWithSearch(requestCommunitySearch,page);
     }
